requirements.txt
matplotlib
tensorflow
flask
numpy
pillow
